Gumroad_Kit/requirements.txt
django==5.2.5
moviepy
elevenlabs
python-dotenv
edge-tts==7.2.5
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pillow
requests
